requirements.txt
rich>=13.7.0
anthropic>=0.25.0
python-dotenv>=1.0.0

